Extract action buttons
